move out registry utils to fix model not registered
